PMM-10029 Fix uint32 conversion.

+++ managed/services/supervisord/supervisord.go
@@ -375,8 +375,8 @@ func (s *Service) UpdateLog(offset uint32) ([]string, uint32, error) {
 		line, err := reader.ReadString('\n')
 		if err == nil {
 			newOffset += uint32(len(line))
-                        if newOffset - offset > s.gRPCMessageMaxSize {
-				return lines, newOffset - len(line), nil
+			if newOffset-offset > s.gRPCMessageMaxSize {
+				return lines, newOffset - uint32(len(line)), nil
 			}
 			lines = append(lines, strings.TrimSuffix(line, "\n"))
 			continue
